+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notible changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),  
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.2]
+
+### Added
+
+- Added the option to double press your drop key to drop everything in your inventory instead of just
+pressing once on an empty slot (**This is on by default**).
+- Added the option to change the time window in which the second press needs to occurr for it to count
+as a double press.
+
 ## [1.1.1]
 
 ### Fixed